feat: separate AbstractServiceBus from CAL layer (closes #72)
