+++ rcal/examples/basic.rs
@@ -8,7 +8,7 @@
 
 use std::sync::Arc;
 
-use rcal::asb;
+use rcal::cal;
 use slog::{error, info, warn};
 
 #[rcal_macros::rcal_main(config = "examples/CALConfig.toml")]
@@ -18,7 +18,7 @@ async fn main() {
     // `rcal_config` and `root_logger` are injected by the macro.
     let config = Arc::new(rcal_config);
 
-    let bus = match asb::get_asb("basic-example", "default", config, root_logger.clone()).await {
+    let bus = match cal::get_cal("basic-example", "default", config, root_logger.clone()).await {
         Ok(b) => b,
         Err(e) => {
             error!(root_logger, "fatal: failed to create ASB"; "error" => %e);
