+++ rcal/examples/custom_tokio.rs
@@ -9,7 +9,7 @@
 
 use std::sync::Arc;
 
-use rcal::asb;
+use rcal::cal;
 use slog::{error, info, warn};
 
 #[rcal_macros::rcal_main(
@@ -25,7 +25,7 @@ async fn main() {
     // `rcal_config` and `my_logger` are injected by the macro.
     let config = Arc::new(rcal_config);
 
-    let bus = match asb::get_asb("custom-example", "default", config, my_logger.clone()).await {
+    let bus = match cal::get_cal("custom-example", "default", config, my_logger.clone()).await {
         Ok(b) => b,
         Err(e) => {
             error!(my_logger, "fatal: failed to create ASB"; "error" => %e);
