+++ rcal/examples/simple_two_service/src/main.rs
@@ -18,7 +18,7 @@ use std::time::Duration;
 
 use slog::{error, info, o, warn};
 
-use rcal::asb::TopicQos;
+use rcal::cal::TopicQos;
 use rcal::asb::zmq::ZmqAsb;
 use rcal::service::{AbstractService, AbstractServiceImpl};
 use rcal::uci::types::*;
