+++ rcal/src/asb/mod.rs
@@ -1,9 +1,12 @@
-//! Abstract Service Bus (ASB) interface.
+//! Abstract Service Bus (ASB) interface — pure transport layer.
 //!
 //! Provides:
 //!   - Service identity and UUID retrieval (CERT CAL-005203)
-//!   - CAL instance lifecycle management (CERT CAL-005201, CAL-005202)
-//!   - ASB connection status – polling **and** callback (§5.9, CERT CAL-016366)
+//!   - ASB connection state machine and status (§5.9)
+//!   - ASB connection lifecycle and status callbacks (CERT CAL-016366)
+//!
+//! Application-facing concerns (message creation, typed writers/readers, QoS)
+//! live in [`crate::cal`].
 //!
 //! ## Specification references
 //! - OMSC-SPC-001 Rev L §5.3, §5.9, Table 5.9-1/2, Figure 5.9-1/2
@@ -12,27 +15,16 @@
 #![allow(dead_code)]
 #![warn(missing_docs)]
 
-use crate::calconfig::CalConfig;
-use crate::uci::CalMessage;
 use crate::uci::base::UUID;
-use crate::uci::types::{
-    ClassificationEnum, ID_Type as _, MessageModeEnum, OwnerProducerChoiceType_,
-};
 use crate::uci::{CalError, CalErrorKind, CalResult};
-use chrono::Utc;
-use lazy_static::lazy_static;
-use std::collections::HashMap;
 use std::env;
 use std::fmt;
 use std::path::Path;
 use std::str::FromStr;
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
+use std::sync::Arc;
 
 #[cfg(feature = "zmq")]
 pub mod zmq;
-#[cfg(feature = "zmq")]
-use zmq::{ZMQ_ASB_ID, ZmqAsb};
 
 // ════════════════════════════════════════════════════════════════════════════
 // Config helpers
@@ -264,17 +256,17 @@ pub trait AsbStatusListener: Send + Sync {
 // AbstractServiceBus
 // ════════════════════════════════════════════════════════════════════════════
 
-/// Central CAL instance interface — Rust equivalent of the C++
-/// `AbstractServiceBusConnection` abstract class.
+/// Pure transport-layer CAL instance interface (object-safe).
 ///
-/// Each instance is bound to a unique (Service Identifier, ASB Identifier)
-/// pair (CERT CAL-005202) and is obtained via [`get_asb`] (CERT CAL-005201).
+/// Covers identity (service/system UUIDs — CERT CAL-005203), connection status
+/// polling and callback (§5.9, CERT CAL-016366), and lifecycle (close()).
+/// Has no knowledge of message types.
 ///
-/// # Responsibilities
+/// Application-facing concerns (message creation, typed writers/readers, QoS)
+/// are in [`crate::cal::AbstractCal`] which extends this trait.
 ///
-/// 1. **Identity** — expose Service and system UUIDs (CERT CAL-005203).
-/// 2. **Connection Status** — polling *and* callback interfaces (§5.9).
-/// 3. **Lifecycle** — initialisation and graceful shutdown.
+/// # CERT coverage
+/// CAL-005201, CAL-005202, CAL-005203, CAL-016366
 pub trait AbstractServiceBus: Send + Sync {
     /// Returns the [`slog::Logger`] associated with this ASB instance.
     fn get_logger(&self) -> &slog::Logger;
@@ -334,16 +326,10 @@ pub trait AbstractServiceBus: Send + Sync {
 
     /// Version string identifying this CAL implementation.
     ///
-    /// Defaults to `<package>/<version>` at build time; override by setting
-    /// the `RCAL_ASB_CONNECTION_VERSION` environment variable during the build.
-    ///
     /// CERT CXX-011176 (`getAbstractServiceBusConnectionVersion()`).
     fn get_asb_connection_version(&self) -> &str;
 
     /// Version string identifying the OMS API against which this CAL was built.
-    ///
-    /// Defaults to `<package>/<version>` at build time; override by setting
-    /// the `RCAL_OMS_API_VERSION` environment variable during the build.
     ///
     /// CERT CXX-012694 (`getOMSApiVersion()`).
     fn get_oms_api_version(&self) -> &str;
@@ -351,9 +337,6 @@ pub trait AbstractServiceBus: Send + Sync {
     // ── Connection Status — Polling ───────────────────────────────────────
 
     /// Returns the current ASB connection status (polling interface, §5.9).
-    ///
-    /// Executes within the caller's thread.  Safe to call in any state,
-    /// including `Failed`.
     fn connection_status(&self) -> &AsbStatus;
 
     // ── Connection Status — Callback ──────────────────────────────────────
@@ -363,8 +346,7 @@ pub trait AbstractServiceBus: Send + Sync {
     /// `on_status_change` is called **immediately** with the current state
     /// (CERT CAL-016366) and subsequently on every state change.
     ///
-    /// Returns `Err(InvalidState)` when called in the `Failed` state
-    /// (Table 5.9-2: `addListener()` in `Failed` → Error).
+    /// Returns `Err(InvalidState)` when called in the `Failed` state.
     fn register_status_listener(&mut self, listener: Arc<dyn AsbStatusListener>) -> CalResult<()>;
 
     /// Unregisters a previously registered ASB status listener.
@@ -379,439 +361,7 @@ pub trait AbstractServiceBus: Send + Sync {
     // ── Lifecycle ─────────────────────────────────────────────────────────
 
     /// Shuts down the CAL instance and releases all associated resources.
-    ///
-    /// After `close()` returns, all Writers and Readers are invalidated,
-    /// registered listeners will not receive further callbacks, and blocked
-    /// `read()` calls are released with an error.
     fn close(&mut self) -> CalResult<()>;
-
-    // ── Message header defaults ───────────────────────────────────────────
-
-    /// Returns the defaults used to pre-populate `MessageHeader` and
-    /// `SecurityInformation` fields when creating a new message via
-    /// [`AbstractServiceBusCreateMessage::create_message`].
-    fn message_header_defaults(&self) -> MessageHeaderDefaults;
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// MessageHeaderDefaults
-// ════════════════════════════════════════════════════════════════════════════
-
-/// Default values applied to every message created through an ASB.
-///
-/// Sourced from the CAL configuration file (`[system]` and `[service]`
-/// sections) plus the compiled-in schema version string.
-#[derive(Debug, Clone)]
-pub struct MessageHeaderDefaults {
-    /// System UUID for the `MessageHeader.SystemID.UUID` field (required).
-    pub system_id: UUID,
-    /// Optional service UUID for the `MessageHeader.ServiceID.UUID` field.
-    pub service_id: Option<UUID>,
-    /// Optional mission UUID for the `MessageHeader.MissionID.UUID` field.
-    pub mission_id: Option<UUID>,
-    /// Schema version string for `MessageHeader.SchemaVersion`.
-    pub schema_version: String,
-    /// Message mode for `MessageHeader.Mode`.
-    pub mode: MessageModeEnum,
-    /// Classification for `SecurityInformation.Classification`.
-    pub classification: ClassificationEnum,
-    /// OwnerProducer entries for `SecurityInformation.OwnerProducer`.
-    pub owner_producer: Vec<OwnerProducerChoiceType_>,
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// MessageListener
-// ════════════════════════════════════════════════════════════════════════════
-
-/// Callback interface for receiving CAL Messages on a subscribed topic.
-///
-/// Register with [`AbstractReader::add_listener`]. Remove with
-/// [`AbstractReader::remove_listener`] (CERT CAL-005396).
-///
-/// # CAL-005392 — single shared reference
-/// Each registered listener receives the **same** `Arc<M>` exactly once per
-/// received message. The inner `M` is immutable for the full duration of the
-/// handler invocation (CERT CAL-016046).
-///
-/// # Thread safety
-/// The CAL may dispatch `on_message` from an internal receive thread.
-/// Implementations must be `Send + Sync`. If internal mutation is required,
-/// use interior mutability (`Mutex`, `AtomicXxx`, etc.).
-///
-/// # CERT coverage
-/// CAL-005379, CAL-005391, CAL-005392, CAL-005396, CAL-016045, CAL-016046
-pub trait MessageListener<M: CalMessage>: Send + Sync {
-    /// Called once per received message instance (CERT CAL-005392).
-    ///
-    /// Must not block indefinitely — the CAL removes the message from its
-    /// internal buffer only after **all** registered listeners' `on_message`
-    /// calls return (CERT CAL-016045).
-    fn on_message(&self, message: &Arc<M>);
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// AbstractWriter
-// ════════════════════════════════════════════════════════════════════════════
-
-/// A topic-bound CAL message publisher (CERT CAL-005368).
-///
-/// Obtained via [`AbstractServiceBusExt::create_writer`] (CERT CAL-005364).
-///
-/// # Error conditions on `write()`
-/// - `CalErrorKind::TopicUnavailable` — topic connection unavailable (CAL-005369)
-/// - `CalErrorKind::ResourcesUnavailable` — platform resources exhausted (CAL-016043)
-/// - `CalErrorKind::InvalidState` — ASB state prohibits writes (Table 5.9-2)
-///
-/// # CERT coverage
-/// CAL-005364, CAL-005368, CAL-005369, CAL-016043
-pub trait AbstractWriter<M: CalMessage>: Send + Sync {
-    /// The Client Topic string this writer is bound to (CERT CAL-005368).
-    fn topic(&self) -> &str;
-
-    /// Publishes `message` on the bound Client Topic.
-    ///
-    /// Returns `Err(TopicUnavailable)` if the topic is unreachable
-    /// (CERT CAL-005369), `Err(ResourcesUnavailable)` if transport resources
-    /// are exhausted (CERT CAL-016043), or `Err(InvalidState)` if the current
-    /// ASB state prohibits writes (Table 5.9-2).
-    fn write(&mut self, message: &M) -> CalResult<()>;
-
-    /// Shuts down this writer and releases all associated resources.
-    ///
-    /// `Box<Self>` consuming receiver prevents use-after-close at the type
-    /// level, consistent with [`AbstractServiceBus::close`].
-    fn close(self: Box<Self>) -> CalResult<()>;
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// AbstractReader
-// ════════════════════════════════════════════════════════════════════════════
-
-/// A topic-bound CAL message subscriber (CERT CAL-005378).
-///
-/// Obtained via [`AbstractServiceBusExt::create_reader`] (CERT CAL-005374).
-/// The topic connection and message buffering are established at creation
-/// time, before this call returns (CERT CAL-005394, CAL-016044).
-///
-/// # Connection timing (stream-based transports)
-///
-/// For transports that use an asynchronous TCP handshake (e.g. ZeroMQ
-/// RADIO/DISH over TCP), the underlying socket connects in the background.
-/// Messages published in the brief window between `create_reader` returning
-/// and the handshake completing may be missed. Callers requiring reliable
-/// first-message delivery should insert a short delay after creation.
-///
-/// # Callback vs polling — mutually exclusive (CAL-016050)
-///
-/// **Callback mode**: register one or more [`MessageListener`]s via
-/// `add_listener`. Each received message is dispatched to all listeners
-/// exactly once, then removed from the buffer (CAL-016045).
-/// Calling `read` or `read_no_wait` while any listener is registered returns
-/// `Err(OperationNotPermitted)` (CAL-016050).
-///
-/// **Polling mode**: call `read` or `read_no_wait` with no listeners
-/// registered. Each call removes the message from the buffer (CAL-016052).
-///
-/// # CERT coverage
-/// CAL-005374, CAL-005378, CAL-005379, CAL-005380, CAL-005391, CAL-005392,
-/// CAL-005394, CAL-005396, CAL-016044, CAL-016045, CAL-016046, CAL-016049,
-/// CAL-016050, CAL-016052
-pub trait AbstractReader<M: CalMessage>: Send + Sync {
-    /// The Client Topic string this reader is bound to (CERT CAL-005378).
-    fn topic(&self) -> &str;
-
-    // ── Callback interface ────────────────────────────────────────────────
-
-    /// Registers a message listener (CERT CAL-005391).
-    ///
-    /// Zero or more listeners may be registered. Each received message
-    /// is dispatched to every listener exactly once (CERT CAL-005392).
-    /// Once any listener is registered, `read` and `read_no_wait` return
-    /// `Err(OperationNotPermitted)` (CERT CAL-016050).
-    fn add_listener(&mut self, listener: Arc<dyn MessageListener<M>>) -> CalResult<()>;
-
-    /// Unregisters a previously registered listener (CERT CAL-005396).
-    ///
-    /// Identified by `Arc` pointer equality. No-op if not registered.
-    fn remove_listener(&mut self, listener: &Arc<dyn MessageListener<M>>) -> CalResult<()>;
-
-    // ── Polling interface ─────────────────────────────────────────────────
-
-    /// Blocking read — waits until a message arrives or the call is released.
-    ///
-    /// Blocks until:
-    /// 1. A message arrives — returns `Ok(Arc<M>)`, message removed from buffer (CAL-016052).
-    /// 2. `timeout` elapses — returns `Ok(None)`.
-    /// 3. This reader is closed — returns `Err(InvalidState)`.
-    /// 4. The ASB enters `Failed` — returns `Err(AsbFailed)`.
-    ///
-    /// `timeout: None` blocks indefinitely until a message or a close event.
-    ///
-    /// Returns `Err(OperationNotPermitted)` if any listener is registered
-    /// (CERT CAL-016050).
-    ///
-    /// # CERT coverage
-    /// CAL-016049, CAL-016050, CAL-016052
-    fn read(&mut self, timeout: Option<Duration>) -> CalResult<Option<Arc<M>>>;
-
-    /// Non-blocking read — returns a buffered message if one is available.
-    ///
-    /// - `Ok(Some(msg))` — message available; removed from buffer (CAL-016052).
-    /// - `Ok(None)` — buffer empty.
-    /// - `Err(OperationNotPermitted)` — listeners registered (CAL-016050).
-    /// - `Err(InvalidState)` — ASB state prohibits reads (Table 5.9-2).
-    /// - `Err(AsbFailed)` — ASB has permanently failed.
-    ///
-    /// # CERT coverage
-    /// CAL-005380, CAL-016050, CAL-016052
-    fn read_no_wait(&mut self) -> CalResult<Option<Arc<M>>>;
-
-    // ── Lifecycle ─────────────────────────────────────────────────────────
-
-    /// Shuts down this reader and releases all associated resources.
-    ///
-    /// After `close()` returns, registered listeners receive no further
-    /// callbacks and any blocked `read` call returns `Err(InvalidState)`.
-    ///
-    /// `Box<Self>` consuming receiver prevents use-after-close.
-    fn close(self: Box<Self>) -> CalResult<()>;
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// QoS settings
-// ════════════════════════════════════════════════════════════════════════════
-
-/// Reliability policy for a Client Topic (CERT CAL-005434, CAL-016076).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum Reliability {
-    /// Best-effort delivery — messages may be dropped (default).
-    #[default]
-    BestEffort,
-    /// Reliable delivery — unacknowledged messages are retransmitted in order.
-    Reliable,
-}
-
-/// Minimum inter-arrival gap for accepted messages (CERT CAL-005431).
-///
-/// The reader silently drops any message received within `min_separation` of
-/// the previously accepted message.
-#[derive(Debug, Clone)]
-pub struct TimeBasedFilter {
-    /// Minimum time that must elapse between two consecutively accepted messages.
-    pub min_separation: Duration,
-}
-
-/// Maximum lifetime for a buffered message (CERT CAL-005437).
-///
-/// Messages older than `max_age` are removed from the receive buffer.
-#[derive(Debug, Clone)]
-pub struct Expiration {
-    /// Age after which a buffered message is discarded.
-    pub max_age: Duration,
-}
-
-/// Bounded message buffer (CERT CAL-005444, CAL-005445, CAL-015746, CAL-016079).
-///
-/// Used for both the writer-side send buffer (`TopicQos::writer_buffer`) and
-/// the reader-side receive buffer (`TopicQos::reader_buffer`). When the count
-/// of buffered messages exceeds `max_messages`, the oldest is dropped.
-#[derive(Debug, Clone)]
-pub struct MessageBuffer {
-    /// Maximum number of messages held in the buffer before the oldest is dropped.
-    pub max_messages: usize,
-}
-
-/// Aggregate Quality of Service settings for a Client Topic (CERT CAL-005210).
-///
-/// Pass to [`AbstractServiceBusExt::create_writer`] or
-/// [`AbstractServiceBusExt::create_reader`] at creation time. The default
-/// value selects best-effort reliability with no filtering, no expiration, and
-/// unbounded buffers.
-#[derive(Debug, Clone, Default)]
-pub struct TopicQos {
-    /// Delivery reliability policy (default: `BestEffort`).
-    pub reliability: Reliability,
-    /// Time-based filter applied on the reader side; `None` disables filtering.
-    pub time_based_filter: Option<TimeBasedFilter>,
-    /// Message lifetime on the reader's receive buffer; `None` disables expiry.
-    pub expiration: Option<Expiration>,
-    /// Writer-side send buffer limit; `None` means unbounded.
-    pub writer_buffer: Option<MessageBuffer>,
-    /// Reader-side receive buffer limit; `None` means unbounded.
-    pub reader_buffer: Option<MessageBuffer>,
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// AbstractServiceBusExt
-// ════════════════════════════════════════════════════════════════════════════
-
-/// Extension trait adding typed factory methods to [`AbstractServiceBus`].
-///
-/// Separated so that `AbstractServiceBus` remains object-safe —
-/// `dyn AbstractServiceBus` can still be stored in `Arc<Mutex<...>>`.
-/// Factory methods are called through a concrete or generic reference.
-///
-/// # CERT coverage
-/// CAL-005364 (`create_writer`), CAL-005374 (`create_reader`)
-pub trait AbstractServiceBusExt<M: CalMessage>: AbstractServiceBus {
-    /// Creates a [`AbstractWriter`] bound to `topic` with the given QoS
-    /// settings (CERT CAL-005364, CAL-005210).
-    ///
-    /// Returns `Err(TopicUnavailable)` if `topic` is not a valid Client Topic
-    /// for this service (CERT CAL-005368, CAL-005369).
-    fn create_writer(
-        &mut self,
-        topic: &str,
-        qos: TopicQos,
-    ) -> CalResult<Box<dyn AbstractWriter<M>>>;
-
-    /// Creates an [`AbstractReader`] bound to `topic` with the given QoS
-    /// settings (CERT CAL-005374, CAL-005210).
-    ///
-    /// The topic connection and message buffering are established before this
-    /// returns (CERT CAL-005394, CAL-016044).
-    ///
-    /// Returns `Err(TopicUnavailable)` if `topic` is not a valid Client Topic
-    /// (CERT CAL-005378).
-    fn create_reader(
-        &mut self,
-        topic: &str,
-        qos: TopicQos,
-    ) -> CalResult<Box<dyn AbstractReader<M>>>;
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// AbstractServiceBusCreateMessage
-// ════════════════════════════════════════════════════════════════════════════
-
-/// Extension trait that adds typed message creation to any [`AbstractServiceBus`].
-///
-/// Implemented as a blanket impl over all `AbstractServiceBus` types so that
-/// transport implementations do not need to duplicate this logic.
-///
-/// Use this instead of constructing message types directly — generated structs
-/// have no public constructors (CERT CAL-016035).
-pub trait AbstractServiceBusCreateMessage {
-    /// Creates a default-initialised instance of message type `M`.
-    ///
-    /// Equivalent to the C++ `<Type>::create(asb)` static factory.
-    fn create_message<M: CalMessage>(&self) -> CalResult<M>;
-}
-
-impl<T: AbstractServiceBus> AbstractServiceBusCreateMessage for T {
-    fn create_message<M: CalMessage>(&self) -> CalResult<M> {
-        let mut msg = M::cal_create();
-        if let Some(mt) = msg.as_message_type_mut() {
-            let defaults = self.message_header_defaults();
-            let hdr = mt.message_header_mut();
-            *hdr.system_id_mut().uuid_mut() = defaults.system_id;
-            *hdr.schema_version_mut() = defaults.schema_version;
-            *hdr.mode_mut() = defaults.mode;
-            *hdr.timestamp_mut() = Utc::now().into();
-            // Optional fields can only be set if already initialized (trait returns Option<&mut>).
-            // service_id and mission_id are set here only if already Some in the message.
-            if let (Some(sid), Some(sfield)) = (defaults.service_id, hdr.service_id_mut()) {
-                *sfield.uuid_mut() = sid;
-            }
-            if let (Some(mid), Some(mfield)) = (defaults.mission_id, hdr.mission_id_mut()) {
-                *mfield.uuid_mut() = mid;
-            }
-            let sec = mt.security_information_mut();
-            *sec.classification_mut() = defaults.classification;
-            let op = sec.owner_producer_mut();
-            op.clear();
-            op.extend(defaults.owner_producer);
-        }
-        Ok(msg)
-    }
-}
-
-// ════════════════════════════════════════════════════════════════════════════
-// Factory
-// ════════════════════════════════════════════════════════════════════════════
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct AsbKey {
-    service_identifier: String,
-    asb_identifier: String,
-}
-
-type AsbInstance = Arc<Mutex<dyn AbstractServiceBus>>;
-type AsbFactoryMap = HashMap<AsbKey, AsbInstance>;
-
-lazy_static! {
-    static ref ASB_FACTORY: tokio::sync::Mutex<AsbFactoryMap> =
-        tokio::sync::Mutex::new(AsbFactoryMap::new());
-}
-
-/// Returns the [`AbstractServiceBus`] instance for `(service_identifier,
-/// asb_identifier)`, creating it if it does not yet exist.
-///
-/// Satisfies CERT CAL-005201 (mechanism to obtain a fully initialised CAL
-/// instance) and CERT CAL-005202 (one instance per unique key pair).
-///
-/// Returns `Err(InitializationFailure)` when no matching (or default)
-/// transport is configured, or when the underlying constructor fails.
-pub async fn get_asb(
-    service_identifier: impl Into<String>,
-    asb_identifier: impl Into<String>,
-    config: Arc<CalConfig>,
-    logger: slog::Logger,
-) -> CalResult<AsbInstance> {
-    let key = AsbKey {
-        service_identifier: service_identifier.into(),
-        asb_identifier: asb_identifier.into(),
-    };
-
-    // Hold the factory lock for the entire construction to prevent concurrent
-    // callers from constructing duplicate instances that bind the same socket.
-    let mut map = ASB_FACTORY.lock().await;
-
-    if let Some(existing) = map.get(&key) {
-        return Ok(Arc::clone(existing));
-    }
-
-    let transport = config
-        .get_transport(&key.asb_identifier)
-        .or_else(|| {
-            config
-                .system
-                .default_transport
-                .as_ref()
-                .and_then(|def| config.get_transport(def))
-        })
-        .ok_or_else(|| {
-            CalError::new(
-                CalErrorKind::InitializationFailure,
-                format!(
-                    "No transport configured for '{}' and no default_transport available.",
-                    key.asb_identifier
-                ),
-            )
-        })?;
-
-    let instance: AsbInstance = match transport.type_.as_str() {
-        #[cfg(feature = "zmq")]
-        ZMQ_ASB_ID => Arc::new(Mutex::new(
-            ZmqAsb::new(
-                key.service_identifier.clone(),
-                key.asb_identifier.clone(),
-                logger,
-                Arc::clone(&config),
-                transport,
-            )
-            .await?,
-        )),
-        other => {
-            return Err(CalError::new(
-                CalErrorKind::InitializationFailure,
-                format!("Unknown ASB transport type: '{other}'."),
-            ));
-        }
-    };
-
-    map.insert(key, Arc::clone(&instance));
-    Ok(instance)
 }
 
 // ════════════════════════════════════════════════════════════════════════════
@@ -850,78 +400,15 @@ pub(crate) static NEXT_TEST_PORT: std::sync::atomic::AtomicU16 =
 #[cfg(test)]
 mod trait_object_safety {
     use super::*;
-    use crate::uci::CalMessage;
 
-    struct Ping;
-    impl CalMessage for Ping {
-        fn message_type_name() -> crate::QName {
-            "test.Ping".into()
-        }
-        fn cal_create() -> Self {
-            Self
-        }
-    }
-
-    // Compile-only assertions — fail at definition site if any trait is not object-safe.
+    // Compile-only assertion — fails at definition site if AbstractServiceBus is not object-safe.
     #[allow(dead_code)]
-    type _W = Box<dyn AbstractWriter<Ping>>;
-    #[allow(dead_code)]
-    type _R = Box<dyn AbstractReader<Ping>>;
-    #[allow(dead_code)]
-    type _L = Arc<dyn MessageListener<Ping>>;
+    type _Asb = Arc<std::sync::Mutex<dyn AbstractServiceBus>>;
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::calconfig::{get_test_config_path, parse_config_from_file};
-    use rcal_macros::init_test_logger;
-    use std::sync::atomic::Ordering;
-
-    #[cfg(feature = "zmq")]
-    #[init_test_logger]
-    #[tokio::test]
-    async fn test_asb_factory_same_key_returns_same_instance() {
-        let p1 = super::NEXT_TEST_PORT.fetch_add(1, Ordering::SeqCst);
-        let p2 = super::NEXT_TEST_PORT.fetch_add(1, Ordering::SeqCst);
-        let config = zmq::test_config_on_ports(&[p1, p2]);
-
-        let a = get_asb("test_svc", "TestZmq", Arc::clone(&config), logger.clone())
-            .await
-            .expect("first get_asb must succeed");
-        let b = get_asb("test_svc", "TestZmq", Arc::clone(&config), logger.clone())
-            .await
-            .expect("second get_asb must succeed");
-        let c = get_asb(
-            "test_svc_2",
-            "TestZmq2",
-            Arc::clone(&config),
-            logger.clone(),
-        )
-        .await
-        .expect("different service must succeed");
-
-        // Same (service, asb) key → same Arc (CERT CAL-005202)
-        assert!(Arc::ptr_eq(&a, &b), "same key should return the same Arc");
-        // Different service key → different Arc
-        assert!(
-            !Arc::ptr_eq(&a, &c),
-            "different service key must be a distinct instance"
-        );
-    }
-
-    #[init_test_logger]
-    #[tokio::test]
-    async fn test_asb_factory_unknown_transport_returns_err() {
-        let no_default_config = Arc::new(
-            parse_config_from_file(&get_test_config_path("calconfig_no_default.toml")).unwrap(),
-        );
-        assert!(
-            get_asb("svc", "dummy", no_default_config, logger)
-                .await
-                .is_err()
-        );
-    }
 
     #[tokio::test]
     async fn test_state_display_does_not_recurse() {
